kubeadm: graduate reset phases
