+++ cmd/kubeadm/app/cmd/phases/BUILD
@@ -34,6 +34,7 @@ filegroup(
         ":package-srcs",
         "//cmd/kubeadm/app/cmd/phases/init:all-srcs",
         "//cmd/kubeadm/app/cmd/phases/join:all-srcs",
+        "//cmd/kubeadm/app/cmd/phases/reset:all-srcs",
         "//cmd/kubeadm/app/cmd/phases/workflow:all-srcs",
     ],
     tags = ["automanaged"],
