+++ cmd/kubeadm/app/cmd/phases/reset/removeetcdmember.go
@@ -1,0 +1,96 @@
+/*
+Copyright 2019 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package phases
+
+import (
+	"errors"
+	"fmt"
+	"path/filepath"
+
+	"k8s.io/klog"
+	kubeadmapi "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm"
+	"k8s.io/kubernetes/cmd/kubeadm/app/cmd/options"
+	"k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/workflow"
+	kubeadmconstants "k8s.io/kubernetes/cmd/kubeadm/app/constants"
+	etcdphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/etcd"
+	utilstaticpod "k8s.io/kubernetes/cmd/kubeadm/app/util/staticpod"
+)
+
+// NewRemoveETCDMemberPhase creates a kubeadm workflow phase for remove-etcd-member
+func NewRemoveETCDMemberPhase() workflow.Phase {
+	return workflow.Phase{
+		Name:  "remove-etcd-member",
+		Short: "Remove a local etcd member.",
+		Long:  "Remove a local etcd member for a control plane node.",
+		Run:   runPreflight,
+		InheritFlags: []string{
+			options.KubeconfigPath,
+		},
+	}
+}
+
+func runRemoveETCDMemberPhase(c workflow.RunData) error {
+	r, ok := c.(resetData)
+	if !ok {
+		return errors.New("remove-etcd-member-phase phase invoked with an invalid data struct")
+	}
+	cfg := r.Cfg()
+
+	// Only clear etcd data when using local etcd.
+	klog.V(1).Infoln("[reset] Checking for etcd config")
+	etcdManifestPath := filepath.Join(kubeadmconstants.KubernetesDir, kubeadmconstants.ManifestsSubDirName, "etcd.yaml")
+	etcdDataDir, err := getEtcdDataDir(etcdManifestPath, cfg)
+	if err == nil {
+		r.AddDirsToClean(etcdDataDir)
+		if cfg != nil {
+			if err := etcdphase.RemoveStackedEtcdMemberFromCluster(r.Client(), cfg); err != nil {
+				klog.Warningf("[reset] failed to remove etcd member: %v\n.Please manually remove this etcd member using etcdctl", err)
+			}
+		}
+	} else {
+		fmt.Println("[reset] No etcd config found. Assuming external etcd")
+		fmt.Println("[reset] Please, manually reset etcd to prevent further issues")
+	}
+
+	return nil
+}
+
+func getEtcdDataDir(manifestPath string, cfg *kubeadmapi.InitConfiguration) (string, error) {
+	const etcdVolumeName = "etcd-data"
+	var dataDir string
+
+	if cfg != nil && cfg.Etcd.Local != nil {
+		return cfg.Etcd.Local.DataDir, nil
+	}
+	klog.Warningln("[reset] No kubeadm config, using etcd pod spec to get data directory")
+
+	etcdPod, err := utilstaticpod.ReadStaticPodFromDisk(manifestPath)
+	if err != nil {
+		return "", err
+	}
+
+	for _, volumeMount := range etcdPod.Spec.Volumes {
+		if volumeMount.Name == etcdVolumeName {
+			dataDir = volumeMount.HostPath.Path
+			break
+		}
+	}
+	if dataDir == "" {
+		return dataDir, errors.New("invalid etcd pod manifest")
+	}
+	return dataDir, nil
+}
