+++ cmd/kubeadm/app/cmd/reset.go
@@ -17,16 +17,10 @@ limitations under the License.
 package cmd
 
 import (
-	"bufio"
 	"fmt"
 	"io"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
 
 	"github.com/lithammer/dedent"
-	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
 	"k8s.io/apimachinery/pkg/util/sets"
@@ -36,18 +30,28 @@ import (
 	kubeadmapiv1beta2 "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta2"
 	"k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/validation"
 	"k8s.io/kubernetes/cmd/kubeadm/app/cmd/options"
+	phases "k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/reset"
 	"k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/workflow"
 	cmdutil "k8s.io/kubernetes/cmd/kubeadm/app/cmd/util"
 	kubeadmconstants "k8s.io/kubernetes/cmd/kubeadm/app/constants"
-	etcdphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/etcd"
-	uploadconfig "k8s.io/kubernetes/cmd/kubeadm/app/phases/uploadconfig"
-	"k8s.io/kubernetes/cmd/kubeadm/app/preflight"
 	kubeadmutil "k8s.io/kubernetes/cmd/kubeadm/app/util"
 	configutil "k8s.io/kubernetes/cmd/kubeadm/app/util/config"
 	utilruntime "k8s.io/kubernetes/cmd/kubeadm/app/util/runtime"
-	utilstaticpod "k8s.io/kubernetes/cmd/kubeadm/app/util/staticpod"
-	"k8s.io/kubernetes/pkg/util/initsystem"
-	utilsexec "k8s.io/utils/exec"
+)
+
+var (
+	iptablesCleanupInstructions = dedent.Dedent(`
+		The reset process does not reset or clean up iptables rules or IPVS tables.
+		If you wish to reset iptables, you must do so manually.
+		For example:
+		iptables -F && iptables -t nat -F && iptables -t mangle -F && iptables -X
+
+		If your cluster was setup to utilize IPVS, run ipvsadm --clear (or similar)
+		to reset your system's IPVS tables.
+
+		The reset process does not clean your kubeconfig files and you must remove them manually.
+		Please, check the contents of the $HOME/.kube/config file.
+	`)
 )
 
 // resetOptions defines all the options exposed via flags by kubeadm reset.
@@ -70,6 +74,7 @@ type resetData struct {
 	inputReader           io.Reader
 	outputWriter          io.Writer
 	cfg                   *kubeadmapi.InitConfiguration
+	dirsToClean           []string
 }
 
 // newResetOptions returns a struct ready for being used for creating cmd join flags.
@@ -166,17 +171,23 @@ func NewCmdReset(in io.Reader, out io.Writer, resetOptions *resetOptions) *cobra
 
 			err = resetRunner.Run(args)
 			kubeadmutil.CheckErr(err)
-			// TODO: remove this once we have all phases in place.
-			// the method joinData.Run() itself should be removed too.
+
+			// Then clean contents from the stateful kubelet, etcd and cni directories
 			data := c.(*resetData)
-			kubeadmutil.CheckErr(data.Run())
+			cleanDirs(data)
+
+			// Output help text instructing user how to remove iptables rules
+			fmt.Print(iptablesCleanupInstructions)
 		},
 	}
 
 	AddResetFlags(cmd.Flags(), resetOptions)
 
 	// initialize the workflow runner with the list of phases
-	// TODO: append phases here
+	resetRunner.AppendPhase(phases.NewPreflightPhase())
+	resetRunner.AppendPhase(phases.NewUpdateClusterStatus())
+	resetRunner.AppendPhase(phases.NewRemoveETCDMemberPhase())
+	resetRunner.AppendPhase(phases.NewCleanupNodePhase())
 
 	// sets the data builder function, that will be used by the runner
 	// both when running the entire workflow or single phases
@@ -189,6 +200,14 @@ func NewCmdReset(in io.Reader, out io.Writer, resetOptions *resetOptions) *cobra
 	resetRunner.BindToCommand(cmd)
 
 	return cmd
+}
+
+func cleanDirs(data *resetData) {
+	fmt.Printf("[reset] Deleting contents of stateful directories: %v\n", data.dirsToClean)
+	for _, dir := range data.dirsToClean {
+		klog.V(1).Infof("[reset] Deleting content of %s", dir)
+		phases.CleanDir(dir)
+	}
 }
 
 // Cfg returns the InitConfiguration.
@@ -221,224 +240,14 @@ func (r *resetData) IgnorePreflightErrors() sets.String {
 	return r.ignorePreflightErrors
 }
 
-func (r *resetData) preflight() error {
-	if !r.ForceReset() {
-		fmt.Println("[reset] WARNING: Changes made to this host by 'kubeadm init' or 'kubeadm join' will be reverted.")
-		fmt.Print("[reset] Are you sure you want to proceed? [y/N]: ")
-		s := bufio.NewScanner(r.InputReader())
-		s.Scan()
-		if err := s.Err(); err != nil {
-			return err
-		}
-		if strings.ToLower(s.Text()) != "y" {
-			return errors.New("Aborted reset operation")
-		}
-	}
-
-	fmt.Println("[preflight] Running pre-flight checks")
-	if err := preflight.RunRootCheckOnly(r.IgnorePreflightErrors()); err != nil {
-		return err
-	}
-
-	return nil
+// AddDirsToClean add a list of dirs to the list of dirs that will be removed.
+func (r *resetData) AddDirsToClean(dirs ...string) {
+	r.dirsToClean = append(r.dirsToClean, dirs...)
 }
 
-// Run reverts any changes made to this host by "kubeadm init" or "kubeadm join".
-func (r *resetData) Run() error {
-	var dirsToClean []string
-	cfg := r.Cfg()
-	certsDir := r.CertificatesDir()
-	client := r.Client()
-
-	err := r.preflight()
-	if err != nil {
-		return err
-	}
-
-	// Reset the ClusterStatus for a given control-plane node.
-	if isControlPlane() && cfg != nil {
-		uploadconfig.ResetClusterStatusForNode(cfg.NodeRegistration.Name, client)
-	}
-
-	// Only clear etcd data when using local etcd.
-	klog.V(1).Infoln("[reset] Checking for etcd config")
-	etcdManifestPath := filepath.Join(kubeadmconstants.KubernetesDir, kubeadmconstants.ManifestsSubDirName, "etcd.yaml")
-	etcdDataDir, err := getEtcdDataDir(etcdManifestPath, cfg)
-	if err == nil {
-		dirsToClean = append(dirsToClean, etcdDataDir)
-		if cfg != nil {
-			if err := etcdphase.RemoveStackedEtcdMemberFromCluster(client, cfg); err != nil {
-				klog.Warningf("[reset] failed to remove etcd member: %v\n.Please manually remove this etcd member using etcdctl", err)
-			}
-		}
-	} else {
-		fmt.Println("[reset] No etcd config found. Assuming external etcd")
-		fmt.Println("[reset] Please manually reset etcd to prevent further issues")
-	}
-
-	// Try to stop the kubelet service
-	klog.V(1).Infoln("[reset] Getting init system")
-	initSystem, err := initsystem.GetInitSystem()
-	if err != nil {
-		klog.Warningln("[reset] The kubelet service could not be stopped by kubeadm. Unable to detect a supported init system!")
-		klog.Warningln("[reset] Please ensure kubelet is stopped manually")
-	} else {
-		fmt.Println("[reset] Stopping the kubelet service")
-		if err := initSystem.ServiceStop("kubelet"); err != nil {
-			klog.Warningf("[reset] The kubelet service could not be stopped by kubeadm: [%v]\n", err)
-			klog.Warningln("[reset] Please ensure kubelet is stopped manually")
-		}
-	}
-
-	// Try to unmount mounted directories under kubeadmconstants.KubeletRunDirectory in order to be able to remove the kubeadmconstants.KubeletRunDirectory directory later
-	fmt.Printf("[reset] Unmounting mounted directories in %q\n", kubeadmconstants.KubeletRunDirectory)
-
-	// In case KubeletRunDirectory holds a symbolic link, evaluate it
-	var absoluteKubeletRunDirectory string
-	absoluteKubeletRunDirectory, err = filepath.EvalSymlinks(kubeadmconstants.KubeletRunDirectory)
-	if err != nil {
-		klog.Errorf("[reset] Failed to evaluate the %q directory. Skipping its unmount and cleanup: %v", kubeadmconstants.KubeletRunDirectory, err)
-	} else {
-		// Only unmount mount points which start with "/var/lib/kubelet" or absolute path of symbolic link, and avoid using empty absoluteKubeletRunDirectory
-		umountDirsCmd := fmt.Sprintf("awk '$2 ~ path {print $2}' path=%s/ /proc/mounts | xargs -r umount", absoluteKubeletRunDirectory)
-		klog.V(1).Infof("[reset] Executing command %q", umountDirsCmd)
-		umountOutputBytes, err := exec.Command("sh", "-c", umountDirsCmd).Output()
-		if err != nil {
-			klog.Errorf("[reset] Failed to unmount mounted directories in %s: %s\n", kubeadmconstants.KubeletRunDirectory, string(umountOutputBytes))
-		} else {
-			// Only clean absoluteKubeletRunDirectory if umountDirsCmd passed without error
-			dirsToClean = append(dirsToClean, absoluteKubeletRunDirectory)
-		}
-	}
-
-	klog.V(1).Info("[reset] Removing Kubernetes-managed containers")
-	if err := removeContainers(utilsexec.New(), r.criSocketPath); err != nil {
-		klog.Errorf("[reset] Failed to remove containers: %v", err)
-	}
-
-	dirsToClean = append(dirsToClean, []string{"/etc/cni/net.d", "/var/lib/dockershim", "/var/run/kubernetes"}...)
-
-	// Then clean contents from the stateful kubelet, etcd and cni directories
-	fmt.Printf("[reset] Deleting contents of stateful directories: %v\n", dirsToClean)
-	for _, dir := range dirsToClean {
-		klog.V(1).Infof("[reset] Deleting content of %s", dir)
-		cleanDir(dir)
-	}
-
-	// Remove contents from the config and pki directories
-	klog.V(1).Infoln("[reset] Removing contents from the config and pki directories")
-	if certsDir != kubeadmapiv1beta2.DefaultCertificatesDir {
-		klog.Warningf("[reset] WARNING: Cleaning a non-default certificates directory: %q\n", certsDir)
-	}
-	resetConfigDir(kubeadmconstants.KubernetesDir, certsDir)
-
-	// Output help text instructing user how to remove iptables rules
-	msg := dedent.Dedent(`
-		The reset process does not reset or clean up iptables rules or IPVS tables.
-		If you wish to reset iptables, you must do so manually.
-		For example:
-		iptables -F && iptables -t nat -F && iptables -t mangle -F && iptables -X
-
-		If your cluster was setup to utilize IPVS, run ipvsadm --clear (or similar)
-		to reset your system's IPVS tables.
-
-		The reset process does not clean your kubeconfig files and you must remove them manually.
-		Please, check the contents of the $HOME/.kube/config file.
-	`)
-	fmt.Print(msg)
-
-	return nil
-}
-
-func getEtcdDataDir(manifestPath string, cfg *kubeadmapi.InitConfiguration) (string, error) {
-	const etcdVolumeName = "etcd-data"
-	var dataDir string
-
-	if cfg != nil && cfg.Etcd.Local != nil {
-		return cfg.Etcd.Local.DataDir, nil
-	}
-	klog.Warningln("[reset] No kubeadm config, using etcd pod spec to get data directory")
-
-	etcdPod, err := utilstaticpod.ReadStaticPodFromDisk(manifestPath)
-	if err != nil {
-		return "", err
-	}
-
-	for _, volumeMount := range etcdPod.Spec.Volumes {
-		if volumeMount.Name == etcdVolumeName {
-			dataDir = volumeMount.HostPath.Path
-			break
-		}
-	}
-	if dataDir == "" {
-		return dataDir, errors.New("invalid etcd pod manifest")
-	}
-	return dataDir, nil
-}
-
-func removeContainers(execer utilsexec.Interface, criSocketPath string) error {
-	containerRuntime, err := utilruntime.NewContainerRuntime(execer, criSocketPath)
-	if err != nil {
-		return err
-	}
-	containers, err := containerRuntime.ListKubeContainers()
-	if err != nil {
-		return err
-	}
-	return containerRuntime.RemoveContainers(containers)
-}
-
-// cleanDir removes everything in a directory, but not the directory itself
-func cleanDir(filePath string) error {
-	// If the directory doesn't even exist there's nothing to do, and we do
-	// not consider this an error
-	if _, err := os.Stat(filePath); os.IsNotExist(err) {
-		return nil
-	}
-
-	d, err := os.Open(filePath)
-	if err != nil {
-		return err
-	}
-	defer d.Close()
-	names, err := d.Readdirnames(-1)
-	if err != nil {
-		return err
-	}
-	for _, name := range names {
-		if err = os.RemoveAll(filepath.Join(filePath, name)); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// resetConfigDir is used to cleanup the files kubeadm writes in /etc/kubernetes/.
-func resetConfigDir(configPathDir, pkiPathDir string) {
-	dirsToClean := []string{
-		filepath.Join(configPathDir, kubeadmconstants.ManifestsSubDirName),
-		pkiPathDir,
-	}
-	fmt.Printf("[reset] Deleting contents of config directories: %v\n", dirsToClean)
-	for _, dir := range dirsToClean {
-		if err := cleanDir(dir); err != nil {
-			klog.Errorf("[reset] Failed to remove directory: %q [%v]\n", dir, err)
-		}
-	}
-
-	filesToClean := []string{
-		filepath.Join(configPathDir, kubeadmconstants.AdminKubeConfigFileName),
-		filepath.Join(configPathDir, kubeadmconstants.KubeletKubeConfigFileName),
-		filepath.Join(configPathDir, kubeadmconstants.KubeletBootstrapKubeConfigFileName),
-		filepath.Join(configPathDir, kubeadmconstants.ControllerManagerKubeConfigFileName),
-		filepath.Join(configPathDir, kubeadmconstants.SchedulerKubeConfigFileName),
-	}
-	fmt.Printf("[reset] Deleting files: %v\n", filesToClean)
-	for _, path := range filesToClean {
-		if err := os.RemoveAll(path); err != nil {
-			klog.Errorf("[reset] Failed to remove file: %q [%v]\n", path, err)
-		}
-	}
+// CRISocketPath returns the criSocketPath.
+func (r *resetData) CRISocketPath() string {
+	return r.criSocketPath
 }
 
 func resetDetectCRISocket(cfg *kubeadmapi.InitConfiguration) (string, error) {
@@ -449,14 +258,4 @@ func resetDetectCRISocket(cfg *kubeadmapi.InitConfiguration) (string, error) {
 
 	// if this fails, try to detect it
 	return utilruntime.DetectCRISocket()
-}
-
-// isControlPlane checks if a node is a control-plane node by looking up
-// the kube-apiserver manifest file
-func isControlPlane() bool {
-	filepath := kubeadmconstants.GetStaticPodFilepath(kubeadmconstants.KubeAPIServer, kubeadmconstants.GetStaticPodDirectory())
-	if _, err := os.Stat(filepath); os.IsNotExist(err) {
-		return false
-	}
-	return true
 }
